Add initrd-based flashing scripts
As an alternative to the traditional flashing scripts, we now have the
option to RCM boot into a minimal Linux system that can write out the
firmware itself.  This is done by upstream in
`tools/kernel_flash/initrd_flash` in the BSP package.  Flashing in this
manner also resolved the issue (#33) where we were unable to flash an
Orin NX after it had been initially flashed from the factory.
Initrd-based flashing will likely be necessary if we ever want to have
the flashing script write out a full NixOS system to an NVMe drive.

Unlike the initrd flashing scripts from NVIDIA's BSP, or from OE4T
meta-tegra, our script is explicitly "multi-spec" and includes firmware
for all Jetson variants associated with a given soc/carrierboard combo.
This means we do not need to query the device first then build the
firmware (like is done in the BSP), or just built for a single variant
(like in OE4T).

In the short-term, we'll still recommend using the existing flash
script, and only recommend using initrd-based flashing if you have a
need (like for the Orin NX).

Closes #33 and #35

Tested on Orin AGX, Orin NX, Xavier AGX, and Xavier NX devkits.