Add UEFI capsule update scripts
To determine if the currently running firmware matches the software,
run, `ota-check-firmware`:
```
$ ota-check-firmware
Current firmware version is: 35.2.1
Current software version is: 35.2.1
```

If these versions do not match, you can update your firmware using the
UEFI Capsule update mechanism. The procedure to do so is below:

To build a capsule update file, build the
`config.system.build.devicePkgs.uefiCapsuleUpdate` attribute from your
NixOS build. For the standard devkit configurations supported in this
repository, one could also run (for example),
`nix build .#uefi-capsule-update-xavier-nx-emmc-devkit`.
Unfortunately, due to some limitations from nvidia's scripts, this build
needs to happen on an `x86_64-linux` machine. This will produce a file
that you can scp (no need for `nix copy`) to the device to update.

Once the file is on the device, run:
```
$ sudo ota-apply-capsule-update example.Cap
$ sudo reboot
```
(Assuming `example.Cap` is the file you copied to the device.)
While the device is rebooting, do not disconnect power.  You should be
able to see a progress bar while the update is being applied. The capsule
update works by updating the non-current slot A/B firmware partitions,
and then rebooting into the new slot. So, if the new firmware does not
boot up to UEFI, it should theoretically rollback to the original
firmware.

UEFI capsule updates have been tested working on orin-agx-devkit,
orin-nx-devkit, xavier-agx-devkit, and xavier-nx-emmc.
For an unknown reason, this currently does _not_ work on
xavier-nx-devkit (the non-production module with an SD card instead of
eMMC).