Add UEFI capsule authentication
These changes allow for the EDK2 UEFI to be build with a public key that
can verify capsule updates as well as for each subsequent capsule update
to be built with a corresponding signing key so that update is accepted
by the firmware. Enabling/disabling capsule authentication and
configuring the keys used during builds are exposed via NixOS module
options.