Add support for single-variant flash scripts and preSignCommands
If there is only a single jetson variant, we can produce a flash script
that doesn't need to detect the variant before flashing. This can make
it so we don't need access to signing keys when performing the flashing
step. (The keys could, for instance, be available via
`extra-sandbox-paths` on the Nix builder machine.)

Using this feature does, however, require specifying the variant
(boardsku, etc.) in the NixOS configuration, which is not always
immediately obvious.

Additionally, adds a new option `preSignCommands` which can be used to
set up the environment to enable the signature part to work.