Unset LD_LIBRARY_PATH
This unset the option `hardware.opengl.setLdLibraryPath`. The latter was
set to true, which is usually a red herring, and is an issue for CUDA
compat: even if we properly patch CUDA packages with a path to
cuda_compat/lib, because /run/opengl-driver has been put in
LD_LIBRARY_PATH, the latter takes precedence and the cuda_compat driver
is never actually selected over the default driver.

A comment mentionned that this was required for libGLX_nvidia.so.0, but
an `ldd` after unsetting `LD_LIBRARY_PATH` doesn't seem to show any
unfound dependency.