Fix eval: Rework some option types to not be nullOr
We were implicitly using the "null" value on the som and carrierBoard
options as a way to specify that the values were unset and that
jetpack-nixos should try to maximize compatibility acroos all types
(with caveats that this may not always include all features). Now, we
just have a separate "generic" value that we can use to represent this.
It's easier to work with since we don't need to have "!= null" checks
everywhere.