Add firmware needed for R816{8,9} network devices
The orin-nano and orin-nx SOMs have realtek NICs that take advantage of
firmware that lives in the `linux-firmware` derivation. Instead of
pulling in all of linux-firmware (1GiB+), just pull in what is needed
for the `r8168` kernel module. Note that in Jetpack 35.3.1 the kernel
module was called `r8169` and in Jetpack 35.4.1 the kernel module was
changed to be `r8168`. This differs from mainline linux where the kernel
module is `r8169`.