Further fix for NvOsLibraryLoad function calls
Programs which linked against both multimedia libs and libEGL_nvidia.so
would end up with two versions of NvOsLibraryLoad and one of them would
be used arbitrarily.

Now, we rename the symbols to have unique names to
avoid the conflict.  This required a version of patchelf that is not yet
in upstream nixpkgs.

As an additional note, this was discovered when runnig the
video_cuda_enc sample, and debugged using LD_DEBUG=all + grepping for
NvOsLibraryLoad and seeing that it was coming from the
libnvos_multimedia lib.