Move `devicePkgs` to the `pkgs.nvidia-jetpack` package-set and clear up ambiguities with systems compatible for flash/fuse scripts
Instead of maintaining a separate package-set location (outside of
`pkgs`), we can just use regular overlays to apply our changes that are
device-specific (stored under the `pkgs.nvidia-jetpack` scope). An alias
is added to the old locations (`config.system.build.jetsonDevicePkgs`,
`config.hardware.nvidia-jetpack.devicePkgs`, etc.) with a warning to
indicate that users should just use `pkgs.nvidia-jetpack`.

Also included is clearing up the ambiguities of what systems are
compatible with flash/fuse scripts. NVIDIA makes the decision for us as
to what platforms we can run these tools on (x86_64-linux only), so we
shouldn't allow for any flash/fuse derivations to be built for
aarch64-linux.

The rundown:
- a jetson device nixos config's `hostPlatform` _must_ be `aarch64-linux`
- a flash/fuse script's `hostPlatform` _must_ be `x86_64-linux`

What we were doing before was mixing package-sets willy-nilly without
much control over which hostPlatform we were dealing with (leading to
lots of usage of hardcoded `pkgsAarch64` to force a package-set to
aarch64). This change makes a logical separation of what needs to be
built with an aarch64 hostPlatform package-set vs an x86_64 hostPlatform
package-set.