Don't require using systemd-boot to get automated firmware updates
Using systemd-boot is not a requirement for applying capsule updates. We
can make the firmware update process a systemd service, which gives us
the behavior of running on every `switch-to-configuration` when the
capsule update file changes. We do not apply the firmware update if we
notice that the currently active nixos system (represented by
/run/current-system) is not the same as the system profile that is
persisted to /nix/var/nix/profiles/system. This allows us to retain the
behavior from before this change where a user running `nixos-rebuild
test` or similar will not have a capsule update applied.