Retain linux-firmware basename
Some files in linux-firmware are symlinks, so when we extract them into
a smaller derivation, we should retain the same filename to ensure a
driver needing firmware is able to find the right file. The `install`
tool will dereference the symlink and use the symlink's basename to
ensure we get the right content at the right location.