Set RAMCODE for Orin AGX Industrial variant
In the normal ./flash.sh process for flashing, it saves a
"chip_info.bin_bak" file, which the tegraflash_impl_t234.py file reads
and extracts the ramcode from before flashing.  Since the single-variant
/ flashcmd approach does not have any chip_info.bin_bak saved to use, we
need to explicitly set the RAMCODE. As far as I can tell thus far, it is
simply set to "1" on Industrial variants.