Move libnvidia-ptxjitcompiler library into l4t-cuda
This library is needed for runtime PTX compilation. It's unclear why
NVIDIA has this library in l4t-3d-core and not in l4t-cuda. Even more so
since the cuda compat package has libcuda as well as
libnvidia-ptxjitcompiler in the same package. meta-tegra does a similar
thing where they pull libnvidia-ptxjitcompiler out of l4t-3d-core and
place it in the same package as libcuda.

Additionally, add matrixMul* samples, including matrixMulDynlinkJIT
specifically, which was failing to run previously.

Fixes #184