Move more packages into the overlay
This change adds more packages from `device-pkgs` into the overlay that
contributes to the system's package-set. The packages moved over in this
change are notably ones that have the `requiredSystemFeatures`
attribute. This means that we could eventually remove the
`requiredSystemFeatures` option and let users change the derivation how
they need in their own overlay.