Ensure declarative boot order doesn't influence initrd flash script
Including the DTBO used for configuring the boot order in the initrd
flash script means that the temporary edk2 build used only for flashing
has the possibility of booting from something other than the
kernel/initrd meant for flashing. Excluding this DTBO when performing
the initrd flash method means that this is no longer possible.