Allow `nixos-rebuild boot` to work with automated firmware updates
In 9cea389171151c28dab47dfc1b08401e87aabfa0, we broke the ability for
using `nixos-rebuild boot` to apply capsule updates. This change adds
that ability back. The systemd service is kept as we still want capsule
updates to be applied in cases where a different bootloader is being
used (or perhaps some other niche setup).