fix unused module errors
https://github.com/NixOS/nixpkgs/pull/366004 introduced a change that
warns and errors out when modules are defined and not acutally set for a
particular target.

This patch rectifies that by going back the the previous method of
ignoring those errors and not breaking the build. Commentary is left
inline to explain the situation and possible future solutions.

tested against orin agx target for ghaf.

there are similar solutions being deployed in nixpkgs to work around
this e.g.

https://github.com/NixOS/nixpkgs/pull/375165

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>