uefi-firmware: update version to include unique hash
Currently, the version field we bake into the UEFI firmware is just the
L4T version.  A simple version field does not any of the additional
customizations (such as patches) we or the end-user may do on top of it.
This makes it difficult to determine if the current firmware on a device
includes the customizations expected.

Additionally, since the firmware autoupgrade feature keys off of the
version number, a user has to know to manually apply a capsule update if
they changed something that affects the platform firmware, which is not
always obvious.

This adds a hash to the version field, like: 35.6.1-06cb4270ebfe. Now,
firmware should automatically be upgraded if there is a change that can
affect platform firmware. This gets us a level of "congruence" similar to
NixOS itself.  The downside is that _any_ change which affects the
derivation hashes included in the firmware build will cause an upgrade,
such as minor bumps to nixpkgs which cause mass-rebuilds. These upgrades
can take approximately 4 minutes during reboot.  Overall, I believe the
tradeoff is worth it.

Fixes #275

Co-authored-by: Elliot Berman <eberman@anduril.com>