Ignore existing capsule file
We don't actually care if an existing capsule file is present on the ESP
when we are applying the capsule built with a NixOS configuration. We
actually _do_ want to always apply the capsule tied to the current NixOS
configuration, since it may be needed for booting that configuration's
kernel or other components (e.g. a NixOS config has a Jetpack 6 kernel
and a capsule update to Jetpack 6 firmware, however some other Jetpack
5.x capsule file is already present).