ota-utils: bake expectedBiosVersion in ota-check-firmware
The ota-check-firmware script is called by both firmware-update.service,
as well as by pre-switch commands run as part of systemd-boot or grub
extraInstallCommands.  In the second case, the
/etc/jetson_expected_bios_version we are referencing will correspond to
the currently running NixOS system, _not_ the one we are intending to
switch to. Thus, the firmware update would not be applied by the
pre-switch commands.  Instead, we now bake this expected firmware
version into the ota-check-firmware script so that it corresponds to the
firmware version we are intending to update to.