kernels: r36: Fix module compilation for nvdisplay
We set KERNEL_OUTPUT when calling the top level makefile to point to the
base kernel derivation. This is expected/prescribed behavior for the top
level makefile in Linux_for_Tegra/source/Makefile. This value, however,
gets propagated into the display drivers. The display drivers depend on
the other OOT modules and the top level makefile attempts to override
this by setting SYSSRC and SYSOUT. The display driver Makefile attempts
to set KERNEL_OUTPUT from SYSOUT/SYSSRC. This "fails" because it's been
(technically) overridden from the command line at the top level
Makefile. Add the override keyword to ensure KERNEL_OUTPUT and
KERNEL_SOURCES are set appropriately if the SYSOUT/SYSSRC variables are
set.