ota-utils: Introduce ota-abort-capsule-update
If one tries to switch-to-configuration a generation which uses new
firmware version (vNext), doesn't reboot, and then
switch-to-confiugration a generation which matches the current firmware,
the vNext capsule update isn't cleared, but should be. Implement a new
helper script: ota-abort-capsule-update which uninstalls a pending
capsule update. Run the script if the firmware version matches.