Introduce jetsonCheckSpecCompatability pre-switch check
We can check whether the bup for a NixOS configuration matches the
machine we're trying to deploy the NixOS system to. This aims to improve
user experience: if the firmware doesn't support the expected board, the
capsule update will fail and it's not clear why unless you inspect the
serial logs.

The check can be manually skipped by setting the
SKIP_TEGRA_COMPAT_CHECK environment variable.