initrdFlash: Instruct device to create a USB serial port and tee output there
The initrd flash script doesn't report progress back to the flashing
station (host PC). When a carrier board doesn't have a serial port, this
is unfortunate because the user doesn't have any way of knowing whether
the flashing is finished, failed, or succeeded. Fortunately, there's a
USB port we can expose an ACM gadget through and pipe the script's
progress through.

In order to create the USB device, we need to ensure the USB driver
stack is loaded. We inherit from boot.initrd.availableKernelModules
because there are carrier board-specific type-c modules that might be
required.

After setting up the gadget, we connect the input/output to /dev/ttyGS0,
the serial port we set up. We tee output to /dev/console and
/dev/ttyGS0, so if someone does monitor the serial port, they still see
the flashing progress there. This does change behavior for those users
if the flashing script fails: the busybox shell that is launched will
only accept input from /dev/ttyGS0.