initrdFlashScript: Force USB to peripheral mode
Carrier boards are not required to support USB OTG mode. Ensure the mode
of the USB port is peripheral. Deices  must support peripheral mode,
else one would not be able to flash it. This only applies to the
recovery mode boot, and not to the images that are being flashed.