uefi-firmware: r36: Clean up fetching repo sources
Consolidate the sha256 sums and which projects need to be fetched. Apply
patches later (patchedRepos). The goal is to make it easier to maintain
with future JetPack upgrades. The current scheme had you reading a lot
of Nix code to find which SHAs to update. It's still messy, but
hopefully easier-to-maintain messy.