Create flasher-pkgs directory
device-pkgs currently contains packages related to flashing devices.
Rename this directory to "flasher-pkgs" to better reflect that it
contains packages for the flasher, as opposed to the device. Also, move
board-automation and flash-tools to this new directory as both are used
from a "host" flasher.

We could call this directory "host-pkgs" to follow the "host/device"
naming scheme commonly used, but this could be confusing in Nix world
because host refers to "where packages are run": host-pkgs would be a
tautology. "flasher-pkgs" is unambigiously packages for the machine
which flashes devices.