fix(overlay): override cudaPackages scope instead of replacing it
I'd like to be able to use some of the packages that are still coming from the
normal nixpkgs' cudaPackages, like cudnn-frontend, which doesn't seem to be
available in the nvidia-jetpack*.cudaPackages, but do seem to compile on aarch64
and jetson nonetheless.

Just replacing the whole cudaPackages attribute set with that of the jetpack
provided one feels like it might be removing packages it shouldn't. Instead, I
overrideScope the original cudaPackages, and only overlay the packages that
nvidia-jetpack actually provides.

I am unsure if the is the right approach, or if there's maybe a better solution.
But this seems to be working?