tree-wide: re-use as much of upstream as possible
I've updated the package expressions for CUDA-related software in Nixpkgs to support as many releases as possible to support nixos-cuda/cuda-legacy.

This allows us to re-use their package expressions, even if we have to make small tweaks to them.

As a result, for JetPack 6/CUDA 12 and later, we can use upstream's packaging.

For JetPack 5/CUDA 11, we still need to use the debian installers. Thankfully, content-wise they're similar enough to the binary archives NVIDIA releases that we can just re-organize their contents and pass them along to upstream's packaging logic.

Signed-off-by: Connor Baker <ConnorBaker01@gmail.com>
Signed-off-by: Connor Baker <cbaker2@anduril.com>