dlopenoverride: Clear dlopen symbol version
We have to clear the dlopen symbol version because when we rename the
symbol it holds onto the original dlopen's symbol version (GLIBC_x.yy).
This version likely does not match the GLIBC version used to compile
dlopenoverride (stdenv.mkDerivation above). It's generally not safe to
completely ignore the version field here; however, the ABI for dlopen
has been reasonably stable and we aren't currently aware of any special
handling required for vesions of glibc. Clear the symbol version
entirely.

Fixes: e949c6ed4238 ("Add dlopen shim")